+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Push empty object when selection is removed in `SelectWithState`.
+- Feature: Update `focusId` on `Label` component.
+- Fix: Add `line-height` on `searchable` of `Select` component.
 - Fix: Update marge on `FormComposer` component.
 
 ## [16.2.0] - 2017-11-22